// apps/mobile/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
// ...
      users: {
        Row: {
          canton_code: string
          class_size: number | null
          created_at: string
          custom_work_hours: number
          education_level: Database["public"]["Enums"]["education_level"] | null
          first_name: string
          is_mode_dark: boolean
          language: string
          last_name: string
          stat_end_date: string | null
          stat_start_date: string | null
          teacher_relief: number | null
          user_id: string
          workload: number | null
        }
        Insert: {
          canton_code: string
          class_size?: number | null
          created_at?: string
          custom_work_hours?: number
          education_level?:
            | Database["public"]["Enums"]["education_level"]
            | null
          first_name: string
          is_mode_dark: boolean
          language?: string
          last_name: string
          stat_end_date?: string | null
          stat_start_date?: string | null
          teacher_relief?: number | null
          user_id: string
          workload?: number | null
        }
        Update: {
          canton_code?: string
          class_size?: number | null
          created_at?: string
          custom_work_hours?: number
          education_level?:
            | Database["public"]["Enums"]["education_level"]
            | null
          first_name?: string
          is_mode_dark?: boolean
          language?: string
          last_name?: string
          stat_end_date?: string | null
          stat_start_date?: string | null
          teacher_relief?: number | null
          user_id?: string
          workload?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "users_canton_code_fkey"
            columns: ["canton_code"]
            isOneToOne: false
            referencedRelation: "cantons"
            referencedColumns: ["code"]
          },
          {
            foreignKeyName: "users_language_fkey"
            columns: ["language"]
            isOneToOne: false
            referencedRelation: "languages"
            referencedColumns: ["locale"]
          },
        ]
      }
    }
// ...
    Enums: {
      education_level:
        | "kindergarten"
        | "foundation_stage"
        | "lower_primary"
        | "grade_3_4"
        | "middle_primary"
        | "lower_secondary"
        | "special_class"
        | "special_school"
        | "vocational_school"
        | "upper_secondary"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never